created baseLogger containing common funtionality of TFEPLogger and TMBARLogger
