+++ tfep/io/__init__.py
@@ -1,5 +1,4 @@
-from tfep.io.log import TFEPLogger
-from tfep.io.tmbarlog import TMBARLogger
+from tfep.io.log import TFEPLogger, TMBARLogger
 from tfep.io.dataset import (
     DictDataset, MergedDataset, TrajectoryDataset,
     TrajectorySubset, get_subsampled_indices,
