--- tfep/io/tmbarlog.py
@@ -1,718 +1,0 @@
-import os
-import h5py
-import numpy as np
-import torch
-from typing import Optional, List, Tuple, Union
-import json
-from tfep.io.log import TFEPLogger
-
-
-class TMBARLogger(TFEPLogger):
-    """Store and retrieve potential energies and CVs during training/evaluations.
-
-    The user can use this to easily store and retrieve arbitrary per-sample
-    quantities such as potential energies and CV values by epoch, batch, or step.
-
-    .. warning::
-
-        Currently, this class is not multi-process or thread safe.
-
-    Current database format
-    -----------------------
-
-    The data is stored in HDF5 format with one file per epoch/step, containing groups
-    for different state mappings. The data is organized in the following structure:
-
-    ```
-    tfep_logs/
-    ├── train/
-    │   ├── epoch-0.h5
-    │   │   ├── 0_1/  # Mapping from state 0 to state 1
-    │   │   │   ├── potential_energy
-    │   │   │   ├── cv_values
-    │   │   │   └── ...
-    │   │   ├── 0_2/  # Mapping from state 0 to state 2
-    │   │   └── ...
-    │   ├── epoch-1.h5
-    │   └── ...
-    ├── eval/
-    │   ├── step-0.h5
-    │   │   ├── 0_1/
-    │   │   ├── 0_2/
-    │   │   └── ...
-    │   └── ...
-    └── metadata.json
-    ```
-
-    Each HDF5 file contains groups for each state mapping, with datasets for the
-    various quantities. The data is organized hierarchically:
-    - First level: state mapping groups (e.g., "0_1" for mapping from state 0 to state 1)
-    - Second level: datasets (potential_energy, cv_values, etc.)
-
-    Finally, a JSON file is used to store metadata about the experiment such as
-    batch and epoch sizes.
-
-    """
-
-    VERSION = '0.1'
-    METADATA_FILE_NAME = 'metadata.json'
-    INDEX_NAMES = ['trajectory_sample_index', 'dataset_sample_index']
-    MASK_NAME = '__mask'
-
-    def __init__(
-            self,
-            save_dir_path='tfep_logs',
-            data_loader=None,
-            train_subdir_name='train',
-            eval_subdir_name='eval',
-            n_states=2,
-            state_names: Optional[List[str]] = None,
-    ):
-        """Constructor.
-
-        Parameters
-        ----------
-        save_dir_path : str, optional
-            The main directory where to save the training and evaluation data.
-        data_loader : torch.utils.data.DataLoader, optional
-            The data loader used for training wrapping a :class:``tfep.io.dataset.TrajectoryDataset``.
-            This must be passed when a new logger is created as it is used to
-            determine epoch, batch, and trajectory dimensions. If ``save_dir_path``
-            points to an existing logger, then this is ignored.
-        train_subdir_name : str, optional
-            The name of the subdirectory where the training data is stored.
-        eval_subdir_name : str, optional
-            The name of the subdirectory where the evaluation data is stored.
-        n_states : int, optional
-            The number of states in the system. Default is 2.
-        state_names : List[str], optional
-            Names for each state. If None, defaults to ['state_0', 'state_1', ...].
-
-        """
-        self._save_dir_path = os.path.realpath(save_dir_path)
-        self._train_dir_path = os.path.join(save_dir_path, train_subdir_name)
-        self._eval_dir_path = os.path.join(save_dir_path, eval_subdir_name)
-        
-        # Initialize state parameters
-        self._n_states = n_states
-        self._state_names = state_names if state_names is not None else [f'state_{i}' for i in range(n_states)]
-        
-        # Generate all possible state mappings
-        self._state_mappings = []
-        for i in range(n_states):
-            for j in range(n_states):
-                if i != j:  # Exclude self-mappings
-                    self._state_mappings.append((i, j))
-        
-        # This keep track of the currently in-memory training/evaluation data.
-        self._loaded_train_idx = None
-        self._loaded_train_data = None  # Dict[name, Tensor].
-        self._loaded_eval_idx = None
-        self._loaded_eval_data = None  # Dict[name, Tensor].
-
-        # Determine whether this is a new logger or we are resuming. The
-        # metadata file is the last file that is created in __init__().
-        metadata_file_path = os.path.join(save_dir_path, self.METADATA_FILE_NAME)
-        resume = os.path.isfile(metadata_file_path)
-
-        # Load metadata.
-        if resume:
-            self._metadata_from_file(metadata_file_path)
-        elif data_loader is None:
-            raise ValueError("When creating a new logger, 'data_loader' must be passed.")
-        else:
-            self._metadata_from_data(data_loader)
-
-        # Create directory structure.
-        os.makedirs(save_dir_path, exist_ok=True)
-        for dir_path in [self._train_dir_path, self._eval_dir_path]:
-            os.makedirs(dir_path, exist_ok=True)
-
-        # Save metadata. This must be the very last thing to do in __init__ for
-        # resuming to work. No need to write on disk if file is already there.
-        if not resume:
-            self._save_metadata(metadata_file_path)
-
-    @property
-    def batch_size(self):
-        """The batch size of the training dataset."""
-        return self._batch_size
-
-    @property
-    def n_samples_per_epoch(self):
-        """The number of samples per training epoch.
-
-        This may be equal to the dataset size, depending on the value of the
-        ``drop_last`` option in ``DataLoader``.
-
-        """
-        return self._n_samples_per_epoch
-
-    @property
-    def n_batches_per_epoch(self):
-        """the number of batches per training epoch."""
-        return int(np.ceil(self.n_samples_per_epoch / self.batch_size))
-
-    @property
-    def save_dir_path(self):
-        """The path to the main directory where the data is stored."""
-        return self._save_dir_path
-
-    @property
-    def n_states(self):
-        """The number of states in the system."""
-        return self._n_states
-
-    @property
-    def state_names(self):
-        """The list of state names."""
-        return self._state_names
-
-    @property
-    def state_mappings(self):
-        """List of all possible state mappings as tuples (from_state, to_state)."""
-        return self._state_mappings
-
-    def read_eval_tensors(
-            self,
-            names: Optional[List[str]] = None,
-            step_idx: Optional[int] = None,
-            epoch_idx: Optional[int] = None,
-            batch_idx: Optional[int] = None,
-            remove_nans: Union[bool, str] = False,
-            sort_by: Optional[str] = None,
-            as_numpy: bool = False,
-            state_mapping: Optional[Tuple[int, int]] = None,
-    ):
-        """Read the tensors generated with the NN model trained for the given number of epoch/batch/step.
-
-        Either ``step_idx`` or both ``epoch_idx`` and ``batch_idx`` must be passed.
-
-        Parameters
-        ----------
-        names : List[str], optional
-            If given, only the tensors saved with the names in this list are
-            returned. Otherwise, all the saved tensors for this step/epoch/batch
-            are returned.
-        step_idx : int, optional
-            If given, the tensors for this optimization step are returned.
-        epoch_idx : int, optional
-            If given, the tensors for this epoch are returned. If ``step_idx``
-            is passed, this is ignored.
-        batch_idx : int, optional
-            If given together with ``epoch_idx``, the tensors for this epoch/batch
-            are returned. If ``step_idx`` is passed, this is ignored.
-        remove_nans : bool or str, optional
-            If ``True`` only the indices corresponding to non NaN entries are
-            returned. If a string, only the indices corresponding to NaN values
-            of ``tensors[remove_nans]`` are returned.
-        sort_by : str, optional
-            If given, all the returned tensors will be sorted based on the tensor
-            with this name (useful if ``sort_by`` is ``'trajectory_sample_index'``).
-            The new data order is also stored on disk thus subsequent calls without
-            saving new data are guaranteed to follow in the same order.
-        as_numpy : bool, optional
-            If ``True``, the tensors are returned as a numpy array rather than
-            PyTorch ``Tensors``.
-        state_mapping : Tuple[int, int], optional
-            A tuple (from_state, to_state) indicating the state mapping.
-            If None, defaults to the first available mapping.
-
-        Returns
-        -------
-        tensors : Dict[str, torch.Tensor]
-            A dictionary mapping the name of the saved tensors to their values.
-        """
-        # Validate input arguments.
-        step_idx, _, _ = self._validate_indices(
-            step_idx, epoch_idx, batch_idx, need_batch=True)
-
-        # Use default state mapping if none provided
-        if state_mapping is None:
-            state_mapping = self._state_mappings[0]
-        elif state_mapping not in self._state_mappings:
-            raise ValueError(f"Invalid state mapping {state_mapping}. Must be one of {self._state_mappings}")
-
-        # Load in memory the data of this NN.
-        self._load_data(step_idx, data_type='eval', state_mapping=state_mapping)
-
-        # Sort the data if requested.
-        if sort_by is not None:
-            sort_indices = np.argsort(self._loaded_eval_data[sort_by])
-            self._loaded_eval_data = {k: v[sort_indices] for k, v in self._loaded_eval_data.items()}
-
-            # Update the data on disk.
-            self._dump_data(data_type='eval', state_mapping=state_mapping)
-
-        # Remove undesired names.
-        if names is None:
-            data = self._loaded_eval_data
-        else:
-            data = {name: self._loaded_eval_data[name] for name in names}
-
-        # Build NaN mask.
-        mask = self._build_mask(remove_nans, data_type='eval')
-        if mask is not None:
-            data = {k: v[mask] for k, v in data.items()}
-
-        # Convert to tensors.
-        if not as_numpy:
-            data = {k: torch.tensor(v) for k, v in data.items()}
-
-        return data
-
-    def read_train_tensors(
-            self,
-            names: Optional[List[str]] = None,
-            step_idx: Optional[int] = None,
-            epoch_idx: Optional[int] = None,
-            batch_idx: Optional[int] = None,
-            remove_nans: Union[bool, str] = False,
-            as_numpy: bool = False,
-            state_mapping: Optional[Tuple[int, int]] = None,
-    ):
-        """Read the tensors saved with ``save_train_tensors``.
-
-        Parameters
-        ----------
-        names : List[str], optional
-            If given, only the tensors saved with the names in this list are
-            returned. Otherwise, all the saved tensors for this step/epoch/batch
-            are returned.
-        step_idx : int, optional
-            If given, the tensors for this optimization step are returned.
-        epoch_idx : int, optional
-            If given, the tensors for this epoch are returned. If ``step_idx``
-            is passed, this is ignored.
-        batch_idx : int, optional
-            If given together with ``epoch_idx``, the tensors for this epoch/batch
-            are returned. If ``step_idx`` is passed, this is ignored.
-        remove_nans : bool or str, optional
-            If ``True`` only the indices corresponding to non NaN entries are
-            returned. If a string, only the indices corresponding to NaN values
-            of ``tensors[remove_nans]`` are returned.
-        as_numpy : bool, optional
-            If ``True``, the tensors are returned as a numpy array rather than
-            PyTorch ``Tensors``.
-        state_mapping : Tuple[int, int], optional
-            A tuple (from_state, to_state) indicating the state mapping.
-            If None, defaults to the first available mapping.
-
-        Returns
-        -------
-        tensors : Dict[str, torch.Tensor]
-            A dictionary mapping the name of the saved tensors to their values.
-        """
-        # Check input arguments.
-        _, epoch_idx, batch_idx = self._validate_indices(
-            step_idx, epoch_idx, batch_idx, need_batch=False)
-
-        # Use default state mapping if none provided
-        if state_mapping is None:
-            state_mapping = self._state_mappings[0]
-        elif state_mapping not in self._state_mappings:
-            raise ValueError(f"Invalid state mapping {state_mapping}. Must be one of {self._state_mappings}")
-
-        # Load in memory the data of this epoch.
-        self._load_data(epoch_idx, data_type='train', state_mapping=state_mapping)
-
-        # Determine which names have to be returned.
-        if names is None:
-            names = [k for k in self._loaded_train_data.keys() if k != self.MASK_NAME]
-
-        # Build mask.
-        mask = self._build_mask(remove_nans, data_type='train')
-
-        # Read data for the specific epoch/batch.
-        tensors = {}
-        for name in names:
-            if batch_idx is None:
-                # Collect data for entire epoch.
-                tensors[name] = self._loaded_train_data[name][mask]
-            else:
-                # Collect data for a single batch.
-                first = self.batch_size * batch_idx
-                last = first + self.batch_size
-                tensors[name] = self._loaded_train_data[name][first:last][mask[first:last]]
-
-        # Convert from numpy to Tensors.
-        if not as_numpy:
-            tensors = {k: torch.tensor(v) for k, v in tensors.items()}
-
-        return tensors
-
-    def save_eval_tensors(
-            self,
-            tensors: dict,
-            step_idx: Optional[int] = None,
-            epoch_idx: Optional[int] = None,
-            batch_idx: Optional[int] = None,
-            update: bool = False,
-            state_mapping: Optional[Tuple[int, int]] = None,
-    ):
-        """Save the tensors generated with the NN model trained for the given number of epoch/batch/step.
-
-        Either ``step_idx`` or both ``epoch_idx`` and ``batch_idx`` must be passed.
-
-        Currently, saving only some of the tensors already on disk is not supported.
-        In other words, the tensor names that will be saved in the first call to
-        ``save_eval_tensors`` for the given step will have to be in all subsequent
-        calls.
-
-        .. warning::
-
-            By default, no check is performed on writing twice data for the same
-            sample indices and the data is simply appended to the existing one.
-            This check is performed only if  ``update`` is set ``True`` and the
-            existing data is overwritten. This check is based on the tensors
-            named (in order of priority) ``'trajectory_sample_indices'`` and
-            ``'dataset_sample_indices'``. Note that this might be an expensive
-            operation and should not be used if not necessary.
-
-        Parameters
-        ----------
-        tensors : Dict[str, torch.Tensor]
-            A dictionary mapping the name of the saved tensors to their values.
-            All tensors must have shape ``(batch_size,)`` unless only ``epoch_idx``
-            is provided, in which case they must have shape ``(n_samples_per_epoch,)``.
-        step_idx : int or None
-            If given, the tensors for this optimization step are saved.
-        epoch_idx : int or None
-            If given, the tensors for this epoch are saved.
-        batch_idx : int or None
-            If given together with ``epoch_idx``, the tensors for this epoch/batch
-            are saved. Otherwise, the data is assumed to be for the entire epoch.
-        update : bool, optional
-            If ``True``, data points corresponding to already stored sample
-            indices are updated (this slows down the method). If ``False``, this
-            check is not performed and all tensors are simply added to the logger.
-        state_mapping : Tuple[int, int], optional
-            A tuple (from_state, to_state) indicating the state mapping.
-            If None, defaults to the first available mapping.
-        """
-        # Warn the user about missing dataset_sample_index nor trajectory_sample_index.
-        self._warn_if_no_indices(tensors)
-
-        # Validate input arguments.
-        step_idx, _, _ = self._validate_indices(
-            step_idx, epoch_idx, batch_idx, need_batch=True)
-
-        # Use default state mapping if none provided
-        if state_mapping is None:
-            state_mapping = self._state_mappings[0]
-        elif state_mapping not in self._state_mappings:
-            raise ValueError(f"Invalid state mapping {state_mapping}. Must be one of {self._state_mappings}")
-
-        # Load in memory the data of this NN.
-        self._load_data(step_idx, data_type='eval', state_mapping=state_mapping)
-
-        # Make sure all known tensors are updated.
-        if len(self._loaded_eval_data) == 0:
-            names = tensors.keys()
-        else:
-            names = self._loaded_eval_data.keys()
-
-        # Convert everything to numpy arrays.
-        try:
-            tensors = {n: tensors[n].detach().cpu().numpy() for n in names}
-        except KeyError:
-            raise KeyError("'tensors' must include all the following Tensors: " + str(list(names)))
-
-        # Check if we need to update some entries (priority to trajectory, then
-        # dataset indices). If we don't find anything, we append all data.
-        if update:
-            # Find the indices and index name.
-            for index_name in self.INDEX_NAMES:
-                try:
-                    tensor_sample_indices = tensors[index_name]
-                except KeyError:
-                    continue
-
-                # Find the common elements.
-                _, tensor_indices, loaded_indices = np.intersect1d(
-                    tensor_sample_indices, self._loaded_eval_data[index_name],
-                    assume_unique=True, return_indices=True
-                )
-
-                # Update the elements in data and remove them from tensors so
-                # that the remaining ones will get appended.
-                if len(tensor_indices) == 0:
-                    break
-
-                for name in names:
-                    self._loaded_eval_data[name][loaded_indices] = tensors[name][tensor_indices]
-                    tensors[name] = np.delete(tensors[name], tensor_indices)
-
-        # Update all tensors.
-        for name in names:
-            # Convert from tensor to array.
-            try:
-                value = tensors[name]
-            except KeyError:
-                raise KeyError("'tensors' must include a Tensor named " + name)
-
-            # Append to current data.
-            try:
-                current_arr = self._loaded_eval_data[name]
-            except KeyError:
-                self._loaded_eval_data[name] = value
-            else:
-                self._loaded_eval_data[name] = np.concatenate((current_arr, value))
-
-        # Update file on disk.
-        self._dump_data(data_type='eval', state_mapping=state_mapping)
-
-    def save_train_tensors(
-            self,
-            tensors: dict,
-            step_idx: Optional[int] = None,
-            epoch_idx: Optional[int] = None,
-            batch_idx: Optional[int] = None,
-            state_mapping: Optional[Tuple[int, int]] = None,
-    ):
-        """Save the tensors generated during the given epoch/batch/step of training.
-
-        Parameters
-        ----------
-        tensors : Dict[str, torch.Tensor]
-            A dictionary mapping the name of the saved tensors to their values.
-            All tensors must have shape ``(batch_size,)`` unless only ``epoch_idx``
-            is provided, in which case they must have shape ``(n_samples_per_epoch,)``.
-        step_idx : int or None
-            If given, the tensors for this optimization step are saved.
-        epoch_idx : int or None
-            If given, the tensors for this epoch are saved.
-        batch_idx : int or None
-            If given together with ``epoch_idx``, the tensors for this epoch/batch
-            are saved. Otherwise, the data is assumed to be for the entire epoch.
-        state_mapping : Tuple[int, int], optional
-            A tuple (from_state, to_state) indicating the state mapping.
-            If None, defaults to the first available mapping.
-        """
-        # Warn the user about missing dataset_sample_index nor trajectory_sample_index.
-        self._warn_if_no_indices(tensors)
-
-        # Validate input arguments.
-        _, epoch_idx, batch_idx = self._validate_indices(
-            step_idx, epoch_idx, batch_idx, need_batch=False)
-        
-        # Use default state mapping if none provided
-        if state_mapping is None:
-            state_mapping = self._state_mappings[0]
-        elif state_mapping not in self._state_mappings:
-            raise ValueError(f"Invalid state mapping {state_mapping}. Must be one of {self._state_mappings}")
-
-        # Load in memory the data of this epoch.
-        self._load_data(epoch_idx, data_type='train', state_mapping=state_mapping)
-
-        # Update all tensors.
-        mask = self._loaded_train_data[self.MASK_NAME]
-        for name, value in tensors.items():
-            # Convert from tensor to array.
-            value = value.detach().cpu().numpy()
-
-            if batch_idx is None:
-                # Assume data is for the entire epoch.
-                self._loaded_train_data[name] = value
-                mask[:] = True
-            else:
-                # Initialize the tensor if this is the first time we see it.
-                try:
-                    saved_array = self._loaded_train_data[name]
-                except KeyError:
-                    saved_array = np.empty(self.n_samples_per_epoch, dtype=value.dtype)
-                    self._loaded_train_data[name] = saved_array
-
-                # Save the data. This is the first index pointing to the start of the batch.
-                first = self.batch_size*batch_idx
-                saved_array[first:first+len(value)] = value
-                mask[first:first+len(value)] = True
-
-        # Update file on disk.
-        self._dump_data(data_type='train', state_mapping=state_mapping)
-
-
-    def _dump_data(self, data_type: str, state_mapping: Tuple[int, int]):
-        """Dump the currently loaded data to the HDF5 file.
-
-        Parameters
-        ----------
-        data_type : str
-            Either 'train' or 'eval'
-        state_mapping : Tuple[int, int]
-            The state mapping (from_state, to_state)
-        """
-        data_attr = '_loaded_' + data_type + '_data'
-        idx_attr = '_loaded_' + data_type + '_idx'
-        idx = getattr(self, idx_attr)
-        
-        file_path = self._get_data_file_path(data_type, idx)
-        group_name = self._get_group_name(state_mapping)
-        
-        with h5py.File(file_path, 'a') as f:
-            # Create group if it doesn't exist
-            if group_name not in f:
-                group = f.create_group(group_name)
-                # Store state mapping as attributes
-                from_state, to_state = state_mapping
-                group.attrs['from_state'] = from_state
-                group.attrs['to_state'] = to_state
-                group.attrs['from_state_name'] = self._state_names[from_state]
-                group.attrs['to_state_name'] = self._state_names[to_state]
-            else:
-                group = f[group_name]
-            
-            # Store the data
-            for name, data in getattr(self, data_attr).items():
-                if name in group:
-                    del group[name]  # Delete existing dataset if it exists
-                group.create_dataset(name, data=data, compression='gzip')
-
-    def _get_data_file_path(self, data_type: str, idx: int) -> str:
-        """Get the path to the HDF5 file for the given data type and index.
-
-        Parameters
-        ----------
-        data_type : str
-            Either 'train' or 'eval'
-        idx : int
-            The epoch or step index
-
-        Returns
-        -------
-        str
-            The path to the HDF5 file
-        """
-        if data_type == 'eval':
-            file_name = f'step-{idx}.h5'
-            dir_path = self._eval_dir_path
-        else:  # train
-            file_name = f'epoch-{idx}.h5'
-            dir_path = self._train_dir_path
-
-        return os.path.join(dir_path, file_name)
-
-    def _get_group_name(self, state_mapping: Tuple[int, int]) -> str:
-        """Get the HDF5 group name for the given state mapping.
-
-        Parameters
-        ----------
-        state_mapping : Tuple[int, int]
-            The state mapping (from_state, to_state)
-
-        Returns
-        -------
-        str
-            The group name in format "state_X_state_Y"
-        """
-        from_state, to_state = state_mapping
-        return f"{self._state_names[from_state]}_{self._state_names[to_state]}"
-
-    def _load_data(self, idx: int, data_type: str, state_mapping: Tuple[int, int]):
-        """Load/initialize the training/evaluation data in memory.
-
-        Parameters
-        ----------
-        idx : int
-            The index of the data to load
-        data_type : str
-            Either 'train' or 'eval'
-        state_mapping : Tuple[int, int]
-            The state mapping (from_state, to_state)
-        """
-        # Check if we have already loaded the data
-        idx_attr = '_loaded_' + data_type + '_idx'
-        if getattr(self, idx_attr) == idx:
-            return
-
-        # Point to new data
-        setattr(self, idx_attr, idx)
-
-        # Check if there is data in the HDF5 file
-        file_path = self._get_data_file_path(data_type, idx)
-        group_name = self._get_group_name(state_mapping)
-        
-        if os.path.isfile(file_path):
-            with h5py.File(file_path, 'r') as f:
-                if group_name in f:
-                    # Load data from HDF5
-                    data_attr = '_loaded_' + data_type + '_data'
-                    setattr(self, data_attr, {k: v[()] for k, v in f[group_name].items()})
-                    return
-
-        # Initialize new data if file or group doesn't exist
-        if data_type == 'eval':
-            data = {}
-        else:
-            # Training data requires a mask
-            data = {self.MASK_NAME: np.full(self.n_samples_per_epoch, fill_value=False)}
-        setattr(self, '_loaded_' + data_type + '_data', data)
-
-    def _metadata_from_data(self, data_loader):
-        """Load metadata from the DataLoader."""
-        # batch_size and drop_last could be in DataLoader or in a batch sampler.
-        self._batch_size = data_loader.batch_size
-        if self._batch_size is None:
-            self._batch_size = data_loader.batch_sampler.batch_size
-            drop_last = data_loader.batch_sampler.drop_last
-        else:
-            drop_last = data_loader.drop_last
-
-        # Determine number of samples per epoch
-        n_dataset_samples = len(data_loader.dataset)
-        if drop_last:
-            self._n_samples_per_epoch = n_dataset_samples - n_dataset_samples%self._batch_size
-        else:
-            self._n_samples_per_epoch = n_dataset_samples
-
-    def _metadata_from_file(self, file_path):
-        """Load metadata from disk.
-
-        Parameters
-        ----------
-        file_path : str
-            Path to the metadata file.
-        """
-        with open(file_path, 'r') as f:
-            metadata = json.load(f)
-        
-        # Load basic metadata
-        self._batch_size = metadata['batch_size']
-        self._n_samples_per_epoch = metadata['n_samples_per_epoch']
-        
-        # Load state-related metadata
-        self._n_states = metadata.get('n_states', 2)
-        self._state_names = metadata.get('state_names', [f'state_{i}' for i in range(self._n_states)])
-        
-        # Load state mappings
-        state_mappings = metadata.get('state_mappings', [])
-        if state_mappings:
-            # Convert string mappings back to tuples
-            self._state_mappings = [tuple(map(int, mapping.split('_'))) for mapping in state_mappings]
-        else:
-            # Generate default mappings if not found
-            self._state_mappings = []
-            for i in range(self._n_states):
-                for j in range(self._n_states):
-                    if i != j:  # Exclude self-mappings
-                        self._state_mappings.append((i, j))
-
-    def _save_metadata(self, file_path):
-        """Save metadata to disk.
-
-        Parameters
-        ----------
-        file_path : str
-            Path to the metadata file.
-        """
-        metadata = {
-            'batch_size': self.batch_size,
-            'n_samples_per_epoch': self.n_samples_per_epoch,
-            'version': self.VERSION,
-            'n_states': self._n_states,
-            'state_names': self._state_names,
-            'state_mappings': [f"{i}_{j}" for i, j in self._state_mappings]
-        }
-        with open(file_path, 'w') as f:
-            json.dump(metadata, f)
-
